Reload grid page after insert/delete committed, to avoid appearance of empty lines
It's become clear that the combination of 1) remote sorting, 2) paging, and 3) batch update
is going to be an issue for the appearance of the grid, because server-side sorting cannot take into
account uncommitted inserts or deletes on the client.  This may mean immediate processing of changes
is the right move... but then users won't have an opportunity to abandon their edits.  It's a puzzle.